add iterator methods
difficult to test yet since we don't have a way of entering values. I tried reusing the HashMap structs but the inner fields are private and couldn't be expressly set.

I'm still not sure that MultiSet<K, V> is right when I know that V will always be a usize for this class. Ive gone back and forth on the siganture so far. We'll do it live.