change type signature
change the type signature to constrain 'V' to usize. I was having trouble figuring out how to do this before, but took a peek at the advanced collections library Counter module and it's a bit clearer. I think this will save a bunch of code, but I won't simply use that because counting is only half the functonality. I need set operations, too.