remove sketch tests
this module started with a set of tests as a sort of sketch of the api and usage. it was mostly based on multiset modules from other libraries and as this one fills out it's clear that an idiomatic rust implementation should follow some merger of the set and map apis instead of some other language. rust analyzer also mistook the commented out tests as doctests, so this removes some minor annoyance in the dev environment. If i need them I can find them in history.