implement difference()
this was tricky. originally I wanted it to return and iterator of
(element, multiplicity) entries but I could not figure out the ownership
of the multiplicity usize. It needs to be generated from both
`self.iter` and `self.other.get()`. It was a mess and the more errors I
worked out the less it resembled the set.rs impl. I could never quite
get all the errors out, either.

So, instead of the developer-ergo friendly [(e, m)...] pattern we now
have a difference method that spits out the same element multiple times
equal to its multiplicity before moving on to the next element. This
will have to do for now. It's not what I expected to do, and I should
finalize it before we get to 1.0