add expected sums to test_difference
this builds off an established pattern for testing iterators by adding a concise sum of the expected array. Other tests can be rewritten with this to avoid magic numbers.

This does not correct issue #5 and instead introduces a test verifying the current behavior instead of introducing a failing test asserting the correct behavior.