add clone constraint, documentation
in order to fix the outstanding issues with iteration over a set we will need to ensure elements can be cloned. we could have, alternatively, stored a vector of elements instead of a usize and returned each element, or some refcell tomfoolery. For my use cases Clone is not a terrible constrain. It will cause issues with complex elements and I added a note about that in the module docs. In common usage I expect this will only impact users dealing using String elements, as literal strings will now need to be converted to String types before they can be used to manipulate the mset. We might have been able to go with the weaker Copy trait instead of Clone, but that complicated the common use case of String elements. I thought it was better to opt for Clone right away. It's possible we could have gotten away with avoiding applying the constraint on retrieval methods like `get` and `containts` but I'll leave that (possible?) optimization for the future

this also trivially changes some remaining uses of `key` to `element` in both variables and documentation without changing behavior, in order to make the library more internally consistent.