remove "any borrowed form" language
At some point we were using the K: Borrow<Q> idiom from hash/map.rs and hash/set.rs and matching that API, but removed it and left the documentation as it was. Re-adding the borrowed form isn't as easy as changing the signature and it's unclear to me what benefit it gives us so we'll fix issue #15 by removing the erroneous documentation instead of adding back the functionality.