remove should be achievable with borrows
Entry requires ownership. Use removal for borrows.

the entry api doesn't allow for ownership yet (https://internals.rust-lang.org/t/pre-rfc-abandonning-morals-in-the-name-of-performance-the-raw-entry-api/7043) but this manages to make it work with a small performance hit. One test failure to figure out in the single remove doctests but the the doctests for remove_times works and the unit tests for both work so I think I can figure that out before this goes to main