replace `capacity` with `len` in tests
recently rust switched to using hashbrown directly for HashMap, which subtly changed the behavior of `capacity` such that our tests fail in beta and nightly versions of rust. Since capacity is how many can be added without reallocation, it's not exactly what we're trying to assert here anyway. `len` provides an exact number of entries